Created arrays for chartjs
I have created the needed arrays for chart js. Next, we need to finish the year link code and combine it  with the popularity code. This will be slightly difficult, as the vue lifecycle is a relatively large problem. However, after we finish, we will only have the charts to do and incorporation into multiple views, as well as cleaning up the project.