Pros and cons of fixed year code.
I have fixed the year code, however there is a small problem. This entire time I have been working towards having the page function with the Year Category being organized and having an actual label in the select options. However, I could not find a way to make this work. The current code that we have will have to do for now.